@@ -35,43 +35,53 @@
         "automatic deduction"
     ],
     "dependencies": {
-        "decompress-zip": "^0.3.1",
         "glob-to-regexp": "^0.4.1",
-        "locate-java-home": "git+https://github.com/viperproject/locate-java-home.git",
-        "mkdirp": "^0.5.1",
         "node-ipc": "^9.1.4",
-        "request": "^2.88.2",
-        "sound-play": "^1.1.0",
-        "stream-json": "^1.7.1",
+        "rimraf": "^3.0.2",
         "strip-json-comments": "3.0.1",
         "sudo-prompt": "8.2.5",
-        "tree-kill": "^1.2.2",
         "unused-filename": "^2.1.0",
         "viz.js": "^2.1.2",
         "vscode-debugadapter": "^1.34.0",
         "vscode-debugprotocol": "^1.34.0",
         "vscode-languageclient": "^5.2.1",
-        "vscode-languageserver": "^5.2.1",
-        "vscode-test": "^1.3.0",
-        "vscode-uri": "^2.1.1"
+        "vscode-uri": "^3.0.2"
     },
     "devDependencies": {
+        "@types/glob-to-regexp": "^0.4.0",
         "@types/mocha": "^8.0.3",
-        "@types/node": "^14.11.8",
+        "@types/node": "^15.6.1",
+        "@types/rimraf": "^3.0.0",
+        "@types/tmp": "^0.2.0",
         "@types/vscode": "^1.43.0",
+        "@typescript-eslint/eslint-plugin": "^4.26.1",
+        "@typescript-eslint/parser": "^4.26.1",
+        "eslint": "^7.28.0",
+        "eslint-webpack-plugin": "^2.5.4",
         "mocha": "^8.1.3",
-        "typescript": "^4.0.3",
-        "typescript.events": "^7.9.0",
-        "vsce": "^1.87.1",
-        "vscode-debugadapter-testsupport": "^1.34.0"
+        "npm-run-all": "^4.1.5",
+        "tmp": "^0.2.1",
+        "ts-loader": "^9.2.2",
+        "typescript": "^4.2.4",
+        "vsce": "^1.88.0",
+        "vscode-test": "^1.3.0",
+        "webpack": "^5.38.1",
+        "webpack-cli": "^4.0.0"
     },
     "scripts": {
-        "compile": "npm run compileServer && tsc -p ./",
-        "watch": "tsc -w -p ./",
-        "prepublish": "npm run prepublishServer && npm run compile",
-        "test": "node ./out/test/runTest.js",
+        "webpack-production": "webpack --mode production",
+        "webpack-development": "webpack --mode development",
+        "webpack-development-watch": "webpack --mode development --watch",
+        "tsc": "tsc",
+        "clean": "rimraf dist",
         "compileServer": "cd ../server && npm run compile && cd ../client",
-        "prepublishServer": "cd ../server && npm install && npm run compile && cd ../client"
+        "compile": "npm-run-all --sequential clean compileServer webpack-development",
+        "watch": "npm-run-all --sequential clean compileServer webpack-development-watch",
+        "vscode:prepublish": "npm-run-all --sequential clean compileServer webpack-production",
+        "test-compile": "npm-run-all --sequential clean compileServer tsc",
+        "pretest": "npm run test-compile",
+        "test": "node ./dist/test/runTest.js",
+        "package": "vsce package"
     },
     "activationEvents": [
         "onLanguage:viper"
