@@ -20,7 +20,6 @@
 
 import * as path from 'path';
 import * as Mocha from 'mocha';
-import * as glob from 'glob';
 
 export async function run(): Promise<void> {
 	// Create the mocha test
@@ -30,24 +29,13 @@ export async function run(): Promise<void> {
         timeout: 600_000, // ms
         color: true,
 	});
-
-	const testsRoot = path.resolve(__dirname, '..');
-
-	const files: Array<string> = await new Promise((resolve, reject) =>
-        glob(
-            "**/*.test.js",
-            {
-                cwd: testsRoot,
-            },
-            (err, result) => {
-                if (err) reject(err)
-                else resolve(result)
-            }
-        )
-    )
-
-	// Add files to the test suite
-    files.forEach(f => mocha.addFile(path.resolve(testsRoot, f)));
+    
+    // do not look for test suites but only execute the one passed as an environment variable:
+    const testSuiteFilename = process.env["VIPER_IDE_TEST_SUITE"];
+    const testsRoot = path.resolve(__dirname, '..');
+    // resolve the path to the test suite here and not already in runTest.ts due to the problems
+    // mentioned in the comment in `getTestSuiteFilenames()`.
+    mocha.addFile(path.resolve(testsRoot, testSuiteFilename));
 
     const failures: number = await new Promise(resolve => mocha.run(resolve));
 
